refactor(discord): finish transitioning 'discord_request' references to 'discord_adapter'
